Update package version and whisper parameters
Increased the package version from 1.0.18 to 1.0.20 as part of the weekly release cycle. In the whisper.ts file, instated a new set of parameters in an attempt to introduce better log_prob, entropy and word thresholds. Removed unnecessary console.log for cleaner code. This is aimed towards improving whisper.js performance.